scripts/c_cpp_build/pyproject.toml
[tool.poetry]
name = "c_cpp_build"
version = "0.1.0"
description = ""
authors = ["lunarium <claymorealt@op.pl>"]

[tool.poetry.dependencies]
python = "^3.8"
colorama = "^0.4.4"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
